ALl test cases apply new AbstractTest to extract configuration initialisation outside
